requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
plaid==0.1.7
plaid-python==3.9.0
pycodestyle==2.5.0
pyflakes==2.2.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
